Factored out YAML parsing.
