+++ st3/sublime_lib/syntax.py
@@ -5,6 +5,7 @@ from collections import namedtuple
 import plistlib
 
 from ._util.collections import projection
+from ._util.yaml import parse_simple_top_level_keys
 from .resource_path import ResourcePath
 
 
@@ -15,29 +16,9 @@ SyntaxInfo = namedtuple('SyntaxInfo', ['path', 'name', 'scope', 'hidden'])
 SyntaxInfo.__new__.__defaults__ = (None, None, False)  # type: ignore
 
 
-def _parse_yaml_value(value):
-    if value.startswith("'"):
-        return value[1:-1].replace("''", "'")
-    elif value.startswith('"'):
-        # JSON and YAML quotation rules are very similar, if not identical
-        return sublime.decode_value(value)
-    elif value == "true":
-        return True
-    elif value == "false":
-        return False
-    elif value == "null":
-        return None
-    else:
-        # Does not handle numbers because we don't expect any
-        return value
-
-
 def get_yaml_metadata(text):
     return projection(
-        dict(
-            map(_parse_yaml_value, match.groups())
-            for match in re.finditer(r'(?m)^(\S.*?):\s*(.*)\s*$', text)
-        ),
+        parse_simple_top_level_keys(text),
         {'name', 'hidden', 'scope'}
     )
 
@@ -64,8 +45,7 @@ def get_syntax_metadata(path):
 
 
 def list_syntaxes():
-    """
-    Return a list of all loaded syntax definitions.
+    """Return a list of all loaded syntax definitions.
 
     Each item is a :class:`namedtuple` with the following properties:
 
@@ -89,9 +69,7 @@ def list_syntaxes():
 
 
 def get_syntax_for_scope(scope):
-    """
-    Returns the last syntax in load order that matches `scope`.
-    """
+    """Returns the last syntax in load order that matches `scope`."""
     return next((
         syntax.path
         for syntax in reversed(list_syntaxes())
